refactor: unbound bytes buffer size
This allows messages to be read and written to payloads larger than 512
bytes, the original UDP limit, as TCP does not have that limitation
(message size restrictions should be handled in the individual transport
mechanism instead). This allows for future expansion as well, as EDNS
and DNSSEC permit larger payloads as well.